+++ test.js
@@ -5,7 +5,7 @@ const babel = require('@babel/core')
 const { TestDirector } = require('test-director')
 const babelPluginSyntaxHighlight = require('.')
 
-const testDirector = new TestDirector()
+const tests = new TestDirector()
 
 /**
  * Creates a function that validates an error is a Babel syntax error with a
@@ -21,7 +21,7 @@ const annotatedBabelSyntaxErrorValidator = messageRegex => error =>
   // Error message has an annotated source location.
   />.+\d+.+\|.+\^/s.test(error.message)
 
-testDirector.add('Line comment expression statement.', () => {
+tests.add('Line comment expression statement.', () => {
   assert.strictEqual(
     babel.transform('// syntax-highlight graphql\n`scalar Upload`', {
       plugins: [babelPluginSyntaxHighlight]
@@ -30,7 +30,7 @@ testDirector.add('Line comment expression statement.', () => {
   )
 })
 
-testDirector.add('Block comment expression statement.', () => {
+tests.add('Block comment expression statement.', () => {
   assert.strictEqual(
     babel.transform('/* syntax-highlight graphql */ `scalar Upload`', {
       plugins: [babelPluginSyntaxHighlight]
@@ -39,7 +39,7 @@ testDirector.add('Block comment expression statement.', () => {
   )
 })
 
-testDirector.add('Variable declarator.', () => {
+tests.add('Variable declarator.', () => {
   assert.strictEqual(
     babel.transform(
       'const a = /* syntax-highlight graphql */ `scalar Upload`',
@@ -49,7 +49,7 @@ testDirector.add('Variable declarator.', () => {
   )
 })
 
-testDirector.add('Arrow function expression.', () => {
+tests.add('Arrow function expression.', () => {
   assert.strictEqual(
     babel.transform(
       'const a = () => /* syntax-highlight graphql */ `scalar Upload`',
@@ -59,7 +59,7 @@ testDirector.add('Arrow function expression.', () => {
   )
 })
 
-testDirector.add('Object property.', () => {
+tests.add('Object property.', () => {
   assert.strictEqual(
     babel.transform(
       'const a = { b: /* syntax-highlight graphql */ `scalar Upload` }',
@@ -69,7 +69,7 @@ testDirector.add('Object property.', () => {
   )
 })
 
-testDirector.add('Template literal escapes.', () => {
+tests.add('Template literal escapes.', () => {
   assert.strictEqual(
     babel.transform(
       // Use markup for simple output. Backslash escaping of ‘`’, ‘\’ and ‘${’
@@ -81,7 +81,7 @@ testDirector.add('Template literal escapes.', () => {
   )
 })
 
-testDirector.add('Unrelated comment and template literal.', () => {
+tests.add('Unrelated comment and template literal.', () => {
   assert.strictEqual(
     babel.transform('/* GraphQL */ ``', {
       plugins: [babelPluginSyntaxHighlight]
@@ -90,7 +90,7 @@ testDirector.add('Unrelated comment and template literal.', () => {
   )
 })
 
-testDirector.add('Multiple relevant and irrelevant comments.', () => {
+tests.add('Multiple relevant and irrelevant comments.', () => {
   assert.strictEqual(
     babel.transform(
       '/* syntax-highlight graphql */ /* GraphQL */ `scalar Upload`',
@@ -100,7 +100,7 @@ testDirector.add('Multiple relevant and irrelevant comments.', () => {
   )
 })
 
-testDirector.add('Multiple relevant comments.', () => {
+tests.add('Multiple relevant comments.', () => {
   assert.throws(() => {
     babel.transform(
       '/* syntax-highlight css */ /* syntax-highlight graphql */ `scalar Upload`',
@@ -109,7 +109,7 @@ testDirector.add('Multiple relevant comments.', () => {
   }, annotatedBabelSyntaxErrorValidator(/Multiple Prism\.js language names specified\./))
 })
 
-testDirector.add('Comment missing the language name.', () => {
+tests.add('Comment missing the language name.', () => {
   assert.throws(() => {
     babel.transform('/* syntax-highlight */ ``', {
       plugins: [babelPluginSyntaxHighlight]
@@ -117,7 +117,7 @@ testDirector.add('Comment missing the language name.', () => {
   }, annotatedBabelSyntaxErrorValidator(/Missing the Prism\.js language name\./))
 })
 
-testDirector.add('Unavailable Prism.js language name.', () => {
+tests.add('Unavailable Prism.js language name.', () => {
   assert.throws(() => {
     babel.transform('/* syntax-highlight _ */ ``', {
       plugins: [babelPluginSyntaxHighlight]
@@ -125,7 +125,7 @@ testDirector.add('Unavailable Prism.js language name.', () => {
   }, annotatedBabelSyntaxErrorValidator(/`_` isn’t an available Prism\.js language name\./))
 })
 
-testDirector.add('Unsupported template literal expression.', () => {
+tests.add('Unsupported template literal expression.', () => {
   assert.throws(() => {
     babel.transform("/* syntax-highlight graphql */ `${''}`", {
       plugins: [babelPluginSyntaxHighlight]
@@ -133,7 +133,7 @@ testDirector.add('Unsupported template literal expression.', () => {
   }, annotatedBabelSyntaxErrorValidator(/Template literal expressions aren’t supported\./))
 })
 
-testDirector.add('Comment not leading a template literal.', () => {
+tests.add('Comment not leading a template literal.', () => {
   assert.strictEqual(
     babel.transform('/* syntax-highlight graphql */', {
       plugins: [babelPluginSyntaxHighlight]
@@ -142,4 +142,4 @@ testDirector.add('Comment not leading a template literal.', () => {
   )
 })
 
-testDirector.run()
+tests.run()